backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pydantic>=2.0.0
gradio>=4.0.0
python-multipart>=0.0.6
loguru>=0.7.0
numpy>=1.24.0
